ensure all join overloads work as join with respective supplied JoinType
